+++ packages/data-mate/src/builder/ListBuilder.ts
@@ -2,7 +2,7 @@ import { castArray } from '@terascope/utils';
 import { DataTypeFieldConfig } from '@terascope/types';
 import { Builder, BuilderOptions, copyVectorToBuilder } from './Builder';
 import { Vector, VectorType } from '../vector';
-import { isSameFieldConfig } from '../core-utils';
+import { isSameFieldConfig } from '../core';
 import { WritableData } from '../data';
 
 export class ListBuilder<T = unknown> extends Builder<Vector<T>> {
