+++ packages/data-mate/src/builder/types/GeoJSONBuilder.ts
@@ -1,6 +1,6 @@
 import { ESGeoShapeType, GeoShape, GeoShapeType } from '@terascope/types';
 import { getTypeOf, isGeoJSON, toString } from '@terascope/utils';
-import { createObject } from '../../core-utils';
+import { createObject } from '../../core';
 import { WritableData } from '../../data';
 import { VectorType } from '../../vector';
 import { Builder, BuilderOptions } from '../Builder';
