+++ packages/data-mate/src/builder/types/ObjectBuilder.ts
@@ -1,5 +1,5 @@
 import { getTypeOf, isPlainObject, toString } from '@terascope/utils';
-import { createObject } from '../../core-utils';
+import { createObject } from '../../core';
 import { WritableData } from '../../data';
 import { VectorType } from '../../vector';
 import { Builder, BuilderOptions, ValueFromFn } from '../Builder';
