+++ packages/data-mate/src/core/config.ts
