+++ packages/data-mate/src/core/index.ts
