+++ packages/data-mate/src/index.ts
@@ -28,7 +28,7 @@ export * from './aggregation-frame';
 export * from './aggregations';
 export * from './builder';
 export * from './column';
-export * from './core-utils';
+export * from './core';
 export * from './data';
 export * from './data-frame';
 export * from './document-matcher';
